Access PoolDistr from shelley ledger state
This will be just enough to derive a committe.. if I just re-use the
pool key hash as the vote signing key of all pools and derive
verification keys this way (to still skip proper key registration).
