+++ ouroboros-consensus-cardano/src/shelley/Ouroboros/Consensus/Shelley/Ledger/Block.hs
@@ -64,7 +64,6 @@ import qualified Data.ByteString.Lazy as Lazy
 import Data.Coerce (coerce)
 import Data.Typeable (Typeable)
 import GHC.Generics (Generic)
-import LeiosVoting (HasLeiosVoting (..))
 import NoThunks.Class (NoThunks (..))
 import Ouroboros.Consensus.Block
 import Ouroboros.Consensus.HardFork.Combinator
@@ -75,12 +74,10 @@ import Ouroboros.Consensus.Protocol.Abstract
   ( ChainDepState
   , TiebreakerView
   )
-import Ouroboros.Consensus.Protocol.Praos (Praos)
 import Ouroboros.Consensus.Protocol.Praos.Common
   ( PraosTiebreakerView
   )
 import Ouroboros.Consensus.Protocol.Signed (SignedHeader)
-import Ouroboros.Consensus.Protocol.TPraos (TPraos)
 import Ouroboros.Consensus.Shelley.Eras
 import Ouroboros.Consensus.Shelley.Ledger.Query.LegacyPParams
 import Ouroboros.Consensus.Shelley.Protocol.Abstract
@@ -360,23 +357,3 @@ instance ShelleyCompatible proto era => Condense (ShelleyBlock proto era) where
 
 instance ShelleyCompatible proto era => Condense (Header (ShelleyBlock proto era)) where
   condense = show . shelleyHeaderRaw
-
--- * Leios voting
-
--- NOTE: Only Dijkstra has Leios voting right now, all earlier Shelley-based
--- eras will never have a committee and thus no voting should happen.
-
--- REVIEW: Use 'proto' instead of Praos/TPraos?
-
--- TODO: Ledger-level type class EraCommittee? LedgerState era -> Committee
-
-instance HasLeiosVoting (ShelleyBlock (TPraos c) ShelleyEra)
-instance HasLeiosVoting (ShelleyBlock (TPraos c) AllegraEra)
-instance HasLeiosVoting (ShelleyBlock (TPraos c) MaryEra)
-instance HasLeiosVoting (ShelleyBlock (TPraos c) AlonzoEra)
-instance HasLeiosVoting (ShelleyBlock (Praos c) BabbageEra)
-instance HasLeiosVoting (ShelleyBlock (Praos c) ConwayEra)
-
-instance HasLeiosVoting (ShelleyBlock (Praos c) DijkstraEra) where
-  getLeiosCommittee _ =
-    Just (error "FIXME: committee selection from ledger state not implemented")
